Post a timer without overwrite
This adds support to the `.post()` method to only post the timer if it
doesn't already exist.  This is useful for implementing application
logic where the timer must fire after `delay` milliseconds, or sooner if
the timer was already set (the current behaviour is to always overwrite
and extend the timer).